refactor: move XML parsing to backend
