+++ package.json
@@ -19,12 +19,14 @@
     "@mui/icons-material": "^5.15.10",
     "@mui/material": "^5.15.10",
     "@tanstack/react-query": "^5.20.5",
+    "axios": "^1.6.7",
     "htmlparser2": "^9.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@biomejs/biome": "^1.5.3",
+    "@netlify/functions": "^2.6.0",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "@vitejs/plugin-basic-ssl": "^1.1.0",
